refactor: extract application bootstrap module
